move strlen_ae to a .cc
This resolves warning:
  non-constant static local variable in inline function may be different
  in different files.

Also make last_ccsid thread-safe.
