+++ include/zos-char-util.h
@@ -200,73 +200,8 @@ public:
 
 #endif // ifdef __cplusplus
 
-inline unsigned strlen_ae(const unsigned char *str, int *code_page,
-                          unsigned long max_len, int *ambiguous) {
-  static int last_ccsid = 819;
-  static const unsigned char _tab_a[256] __attribute__((aligned(8))) = {
-      1, 1, 1, 1, 1, 1, 1, 0, 0, 0, 0, 0, 0, 0, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1,
-      1, 1, 1, 1, 1, 1, 1, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
-      0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
-      0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
-      0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
-      0, 0, 0, 0, 0, 0, 0, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1,
-      1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1,
-      1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1,
-      1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1,
-      1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1,
-      1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1,
-  };
-  static const unsigned char _tab_e[256] __attribute__((aligned(8))) = {
-      1, 1, 1, 1, 1, 0, 1, 1, 1, 1, 1, 0, 0, 0, 1, 1, 1, 1, 1, 1, 1, 0, 0, 1,
-      1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 0,
-      1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 0, 1, 1, 1, 1, 1, 1, 1,
-      1, 1, 1, 0, 0, 0, 0, 0, 0, 1, 1, 1, 1, 1, 1, 1, 1, 1, 0, 0, 0, 0, 0, 0,
-      0, 0, 1, 1, 1, 1, 1, 1, 1, 1, 1, 0, 0, 0, 0, 0, 1, 1, 1, 1, 1, 1, 1, 1,
-      1, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 1, 1, 1, 1, 1,
-      1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 1, 1, 1, 1, 1, 1, 0, 0, 0, 0, 0, 0, 0,
-      0, 0, 1, 1, 1, 0, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 0, 1, 1,
-      0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 1, 1, 1, 1, 1, 0, 0, 0, 0, 0, 0, 0, 0,
-      0, 0, 1, 1, 1, 1, 1, 1, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 1, 1, 1, 1, 1, 1,
-      0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 1, 1, 1, 1, 1,
-  };
-  unsigned long bytes;
-  unsigned long code_out;
-  const unsigned char *start;
-
-  bytes = max_len;
-  code_out = 0;
-  start = str;
-  __asm volatile(" trte %1,%3,0\n"
-                 " jo *-4\n"
-                 : __ZL_NR("+",r3)(bytes), __ZL_NR("+",r2)(str), "+r"(bytes), "+r"(code_out)
-                 : __ZL_NR("",r1)(_tab_a)
-                 :);
-  unsigned a_len = str - start;
-
-  bytes = max_len;
-  code_out = 0;
-  str = start;
-  __asm volatile(" trte %1,%3,0\n"
-                 " jo *-4\n"
-                 : __ZL_NR("+",r3)(bytes), __ZL_NR("+",r2)(str), "+r"(bytes), "+r"(code_out)
-                 : __ZL_NR("",r1)(_tab_e)
-                 :);
-  unsigned e_len = str - start;
-  if (a_len > e_len) {
-    *code_page = 819;
-    last_ccsid = 819;
-    *ambiguous = 0;
-    return a_len;
-  } else if (e_len > a_len) {
-    *code_page = 1047;
-    last_ccsid = 1047;
-    *ambiguous = 0;
-    return e_len;
-  }
-  *code_page = last_ccsid;
-  *ambiguous = 1;
-  return a_len;
-}
+__Z_EXPORT unsigned strlen_ae(const unsigned char *str, int *code_page,
+                              unsigned long max_len, int *ambiguous);
 
 inline unsigned strlen_e(const unsigned char *str, unsigned size) {
   static const unsigned char _tab_e[256] __attribute__((aligned(8))) = {
