+++ src/zos-char-util.cc
@@ -10,6 +10,7 @@
 #include "zos-char-util.h"
 #include "zos-base.h"
 #include "zos-io.h"
+#include "zos-tls.h"
 
 #include <_Ccsid.h>
 #include <ctest.h>
@@ -578,6 +579,76 @@ void __set_autocvt_on_fd_stream(int fd, unsigned short ccsid,
     fcntl(fd, F_CONTROL_CVT, &req);
     fcntl(fd, F_SETTAG, &tag);
   }
+}
+
+unsigned strlen_ae(const unsigned char *str, int *code_page,
+                   unsigned long max_len, int *ambiguous) {
+  static __tlssim<int> last_ccsid(819);
+  #define last_ccsid (*last_ccsid.access())
+
+  static const unsigned char _tab_a[256] __attribute__((aligned(8))) = {
+      1, 1, 1, 1, 1, 1, 1, 0, 0, 0, 0, 0, 0, 0, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1,
+      1, 1, 1, 1, 1, 1, 1, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
+      0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
+      0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
+      0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
+      0, 0, 0, 0, 0, 0, 0, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1,
+      1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1,
+      1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1,
+      1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1,
+      1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1,
+      1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1,
+  };
+  static const unsigned char _tab_e[256] __attribute__((aligned(8))) = {
+      1, 1, 1, 1, 1, 0, 1, 1, 1, 1, 1, 0, 0, 0, 1, 1, 1, 1, 1, 1, 1, 0, 0, 1,
+      1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 0,
+      1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 0, 1, 1, 1, 1, 1, 1, 1,
+      1, 1, 1, 0, 0, 0, 0, 0, 0, 1, 1, 1, 1, 1, 1, 1, 1, 1, 0, 0, 0, 0, 0, 0,
+      0, 0, 1, 1, 1, 1, 1, 1, 1, 1, 1, 0, 0, 0, 0, 0, 1, 1, 1, 1, 1, 1, 1, 1,
+      1, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 1, 1, 1, 1, 1,
+      1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 1, 1, 1, 1, 1, 1, 0, 0, 0, 0, 0, 0, 0,
+      0, 0, 1, 1, 1, 0, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 0, 1, 1,
+      0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 1, 1, 1, 1, 1, 0, 0, 0, 0, 0, 0, 0, 0,
+      0, 0, 1, 1, 1, 1, 1, 1, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 1, 1, 1, 1, 1, 1,
+      0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 1, 1, 1, 1, 1,
+  };
+  unsigned long bytes;
+  unsigned long code_out;
+  const unsigned char *start;
+
+  bytes = max_len;
+  code_out = 0;
+  start = str;
+  __asm volatile(" trte %1,%3,0\n"
+                 " jo *-4\n"
+                 : __ZL_NR("+",r3)(bytes), __ZL_NR("+",r2)(str), "+r"(bytes), "+r"(code_out)
+                 : __ZL_NR("",r1)(_tab_a)
+                 :);
+  unsigned a_len = str - start;
+
+  bytes = max_len;
+  code_out = 0;
+  str = start;
+  __asm volatile(" trte %1,%3,0\n"
+                 " jo *-4\n"
+                 : __ZL_NR("+",r3)(bytes), __ZL_NR("+",r2)(str), "+r"(bytes), "+r"(code_out)
+                 : __ZL_NR("",r1)(_tab_e)
+                 :);
+  unsigned e_len = str - start;
+  if (a_len > e_len) {
+    *code_page = 819;
+    last_ccsid = 819;
+    *ambiguous = 0;
+    return a_len;
+  } else if (e_len > a_len) {
+    *code_page = 1047;
+    last_ccsid = 1047;
+    *ambiguous = 0;
+    return e_len;
+  }
+  *code_page = last_ccsid;
+  *ambiguous = 1;
+  return a_len;
 }
 
 #ifdef __cplusplus
